Split run_one_test.py to a libraries
Easier to manage and work on those.
