+++ scripts/launcher/run_one_test.py
@@ -32,223 +32,14 @@
 # 77 - Something needed by the test doesn't exist, so skip
 # 99 - Test preparation failed
 
-from argparse import ArgumentParser, RawDescriptionHelpFormatter
-from enum import Enum
-from contextlib import AbstractContextManager
-from tempfile import mkdtemp
-from glob import glob
 
-import os
 import re
 import shutil
 import subprocess
 
-SHELL_INTERFACE_PATH = "launcher_interface.sh"
-
-
-class KeepLevel(Enum):
-    NOTHING = 0
-    LOGS_ONLY = 1
-    EVERYTHING = 2
-
-
-class RunnerConfiguration(object):
-
-    def __init__(self):
-        super().__init__()
-
-        self._parser = ArgumentParser(description="""
-        Run one kickstart test.
-
-        This should be run in parallel by main script. It is not supposed to be invoked manually.
-        """, formatter_class=RawDescriptionHelpFormatter)
-        self._confiure_parser()
-
-        self._ks_test_name = ""
-        self._sh_path = ""
-        self._ks_path = ""
-        self._image_path = ""
-        self._keep_option = KeepLevel.NOTHING
-        self._updates_img_path = ""
-
-    def _confiure_parser(self):
-        self._parser.add_argument("kickstart_test", metavar="KS test controller",
-                                  type=str, help="Kickstart test to run")
-        self._parser.add_argument("image", metavar="Image path", type=str,
-                                  help="Image used to run specified kickstart test")
-        self._parser.add_argument("--keep", '-k', metavar="0,1,2", type=int,
-                                  dest="keep", help="""
-                                  Set the level of what should be kept after a test
-
-                                  Valid potions:
-                                  0 - Remove everything
-                                  1 - Remove disk image and kickstart test
-                                  2 - Keep everything
-                                  """)
-        self._parser.add_argument("--updates", '-u', metavar="Path",
-                                  type=str, dest="updates_path",
-                                  help="Updates image path used in the test")
-
-    @property
-    def shell_test_path(self):
-        return self._sh_path
-
-    @property
-    def ks_test_path(self):
-        return self._ks_path
-
-    @property
-    def ks_test_name(self):
-        return self._ks_test_name
-
-    @property
-    def boot_image(self):
-        return self._image_path
-
-    @property
-    def keep_level(self):
-        return self._keep_option
-
-    @property
-    def update_img_path(self):
-        return self._updates_img_path
-
-    @property
-    def script_path(self):
-        return os.path.dirname(os.path.realpath(__file__))
-
-    def process_argument(self):
-        ns = self._parser.parse_args()
-
-        self._sh_path = ns.kickstart_test
-        self._image_path = ns.image
-
-        base_path = os.path.splitext(self._sh_path)[0]
-        self._ks_path = "{}{}".format(base_path, ".ks")
-
-        self._ks_test_name = os.path.basename(base_path)
-
-        if ns.keep and ns.keep not in [0, 1, 2]:
-            raise ValueError("keep parameter can contain only numbers: 0, 1 or 2 !")
-        elif ns.keep is not None:
-            self._keep_option = KeepLevel(ns.keep)
-
-        if ns.updates_path:
-            self._updates_img_path = ns.updates_path
-
-        self._check_arguments()
-
-    def _check_arguments(self):
-        if not os.path.exists(self._sh_path):
-            raise IOError("Kickstart test shell file '{}' does not exists!".format(self._sh_path))
-        elif not os.path.exists(self._ks_path):
-            raise IOError("Kickstart file '{}' does not exists!".format(self._ks_path))
-        elif not os.path.exists(self._image_path):
-            raise IOError("Boot iso file '{}' does not exists!".format(self._image_path))
-        elif self._updates_img_path and not os.path.exists(self._updates_img_path):
-            raise IOError("Updates image '{}' does not exists!".format(self._updates_img_path))
-
-
-class TempManager(AbstractContextManager):
-
-    def __init__(self, keep_type, test_name):
-        super().__init__()
-
-        self._tmp_dir = None
-        self._keep_type = keep_type
-        self._test_name = test_name
-
-    def __enter__(self):
-        prefix = "kstest-{}.".format(self._test_name)
-        self._tmp_dir = mkdtemp(prefix=prefix, dir="/var/tmp")
-
-        return self._tmp_dir
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        if self._keep_type is KeepLevel.NOTHING:
-            shutil.rmtree(self._tmp_dir)
-        elif self._keep_type is KeepLevel.LOGS_ONLY:
-            images_path = self._tmp_join_path("disk-*.img")
-            ks_path = self._tmp_join_path("*.ks")
-
-            for f in glob(images_path):
-                os.remove(f)
-
-            for f in glob(ks_path):
-                os.remove(f)
-
-    def _tmp_join_path(self, file_path):
-        return os.path.join(self._tmp_dir, file_path)
-
-
-class ShellLauncher(object):
-
-    def __init__(self, configuration, tmp_dir):
-        super().__init__()
-        self._conf = configuration
-        self._tmp_dir = tmp_dir
-
-    def run_prepare(self):
-        ret = self._run_shell_func("prepare")
-        return ret.stdout.decode()
-
-    def run_cleanup(self):
-        ret = self._run_shell_func("cleanup")
-        return ret.stdout.decode()
-
-    def run_prepare_disks(self):
-        ret = self._run_shell_func("prepare_disks")
-        return ret.stdout.decode()
-
-    def run_prepare_network(self):
-        ret = self._run_shell_func("prepare_network")
-        return ret.stdout.decode()
-
-    def run_kernel_args(self):
-        ret = self._run_shell_func("kernel_args")
-        return ret.stdout.decode()
-
-    def run_additional_runner_args(self):
-        ret = self._run_shell_func("additional_runner_args")
-        return ret.stdout.decode()
-
-    def _run_shell_func(self, func_name):
-        cmd_args = []
-        script_path = os.path.join(self._conf.script_path, SHELL_INTERFACE_PATH)
-
-        cmd_args.append(script_path)
-        cmd_args.append("-i")
-        cmd_args.append(self._conf.boot_image)
-        cmd_args.append("-k")
-        cmd_args.append(str(self._conf.keep_level.value))
-
-        if self._conf.update_img_path:
-            cmd_args.append("-u")
-            cmd_args.append(self._conf.update_img_path)
-
-        cmd_args.append("-w")
-        cmd_args.append(self._tmp_dir)
-        cmd_args.append("-t")
-        cmd_args.append(self._conf.shell_test_path)
-
-        cmd_args.append(func_name)
-
-        out = subprocess.run(cmd_args, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-
-        try:
-            out.check_returncode()
-            return out
-        except subprocess.CalledProcessError as e:
-            self._report_error(e)
-            raise e
-
-    @staticmethod
-    def _report_error(exc):
-        print("Failed to run subprocess:")
-        print("stderr:")
-        print(exc.stderr)
-        print("stdout:")
-        print(exc.stdout)
+from lib.temp_manager import TempManager
+from lib.configuration import RunnerConfiguration
+from lib.shell_launcher import ShellLauncher
 
 
 class Runner(object):
